+++ lib/KTable/__tests__/useSorting.spec.js
@@ -6,7 +6,7 @@ import useSorting, {
   DATA_TYPE_NUMERIC,
   DATA_TYPE_DATE,
   DATA_TYPE_OTHERS,
-} from '../';
+} from '../useSorting';
 
 describe('useSorting', () => {
   let headers, rows, defaultSort, disableBuiltinSorting;
